+++ frc971/orin/stag_decoder.h
@@ -39,7 +39,7 @@ class STagDecoderGPUWorker
 public:
     STagDecoderGPUWorker(const std::string &modelPath,
                          const std::string &onnxModelFilename,
-                         const size_t batchSize,
+                         const int32_t batchSize,
                          Timings &timing);
     virtual ~STagDecoderGPUWorker() = default;
     void runInference(std::vector<std::vector<Stage2KeypointGroup>> &stage2KeypointGroups,
